fix(metric): :bug: rename UIED metric from m23 to m25
